+++ Frends.CSV.ConvertToXML/Frends.CSV.ConvertToXML/Definitions/Input.cs
@@ -1,4 +1,5 @@
-﻿using System.ComponentModel;
+﻿using System;
+using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 
 namespace Frends.CSV.ConvertToXML.Definitions;
@@ -19,13 +20,12 @@ public class Input
     /// Delimiter.
     /// </summary>
     /// <example>;</example>
-    [DefaultValue("\";\"")]
-    public string Delimiter { get; set; }
-
+    [DefaultValue(";")]
+    public string Delimiter { get; set; } = ";";
     /// <summary>
     /// You can map columns to specific types. 
     /// The order of the columns are used for mapping, that means that the ColumnSpecification elements need to be created in the same order as the CSV fields.
     /// </summary>
     /// <example>[ { foo, String } ]</example>
-    public ColumnSpecification[] ColumnSpecifications { get; set; }
+    public ColumnSpecification[] ColumnSpecifications { get; set; } = Array.Empty<ColumnSpecification>();
 }